prod/services/pdf-service/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0

# File handling
python-multipart>=0.0.6

# PDF processing dependencies
pdfplumber>=0.9.0
PyMuPDF>=1.23.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0
pytesseract>=0.3.10
Pillow>=9.5.0
numpy>=1.24.0

# HTTP client
httpx>=0.25.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging
structlog>=23.0.0